refactor: split Steel Design into four pages
One page with three tabs became four pages, one per calculator: Steel
Beam, Steel Column, Bolted Connection, Welded Connection. Each gets its
own route, its own printable report, its own docs entry and its own
trial allowance — five free runs each rather than five for all of them
together, which is what the pricing page has been promising a
"single-purpose calculator" all along.

The connection tab was really two calculators behind a Type dropdown, so
splitting on tabs alone would have left it. Bolted keeps the LRFD work —
phiRn per bolt, block shear §J4.3, out-of-plane §J3.7, prying §J3.9 —
and the weld half is gone, because it was a three-row fillet-length
check and /welded-connection already solves the whole eccentric weld
group. The one thing that tab did better was naming the ELECTRODE
instead of asking for a stress, so the weld page now offers E70/E80/
E90/E100 (from the engine's own table, via a new `FEXX_BY_CLASS`, so the
two cannot drift) with Custom for a filler that is not in it.

That makes the connection calc endpoint bolted-only: `electrode`,
`wSize`, `weld` and `weldCapacity` are gone from `ConnectionCalcInput`
and `ConnectionCalcResult`. Nothing else consumed them, and there is no
server implementation of the route to keep in sync — `calcApi` is the
only place it exists.

Shared atoms move out of the page rather than being copied four times:
`components/steelUi.tsx` for the widgets and `lib/steelShapes.ts` for
the W-shape catalogue and grade table. The split between them is not
taste — a module exporting both components and plain values breaks Fast
Refresh, and the lint rule enforces it. `ok()` and `badge()` became
`<Verdict>` and `<ZoneBadge>` for the same reason.

/steel redirects to /steel/beam, classified PUBLIC and dropped from the
trial list like every other stub. The route test that asserted
"/bolted-connection redirects to /steel" is now the general statement —
/steel and /geotech are the stubs, and each of the four steel pages
holds its own allowance and exhausts independently.

Verified: tsc -b clean, lint clean, 3851 tests green. All four pages
driven in headless Chromium: correct heading, real results (φMn, φPn,
φRn + block shear, weld peak force), a worked solution on each, "5 free
runs left" on each, no console errors — and /steel lands on the beam.
