+++ webapp/src/lib/docsContent.ts
@@ -22,9 +22,8 @@ export const DOC_TOOLS: DocTool[] = [
  * page — it only passes for routes deliberately folded into a parent.
  */
 export const DOC_ROUTE_ALIASES: Record<string, string> = {
-  // the standalone bolted-connection page was folded into Steel Design's
-  // Connection tab; the route redirects there and is documented there.
-  '/bolted-connection': 'steel-design',
+  // Steel Design split into four pages; /steel is a stub that lands on the beam
+  '/steel': 'steel-beam',
   // the geotechnical index was navigation only; the route now redirects
   '/geotech': 'bearing-capacity',
   '/signup': 'account', '/forgot-password': 'account', '/reset-password': 'account',
