+++ webapp/src/lib/docsContentDesign.ts
@@ -242,17 +242,19 @@ export const DESIGN_TOOLS: DocTool[] = [
     ],
   },
   {
-    id: 'steel-design',
-    name: 'Steel Design',
-    route: '/steel',
+    id: 'steel-beam',
+    name: 'Steel Beam',
+    route: '/steel/beam',
     group: 'Steel & timber',
-    summary: 'AISC member checks — beam flexure and shear, column compression, combined forces — plus shear tabs and moment connections.',
-    basis: 'AISC 360-16 §F2 flexure, §G2.1 shear, §E3 compression, §H1-1 combined, §J bolts/welds/block shear.',
+    summary: 'Flexure with lateral-torsional buckling, shear, and service deflection for a simply-supported W-section.',
+    basis: 'AISC 360-16 §F2 flexure, §G2.1 shear; deflection against L/360 live and L/240 total.',
     sections: [
       {
-        id: 'sd-beam',
-        title: 'Beam',
+        id: 'sb-in',
+        title: 'Inputs',
         controls: [
+          { kind: 'choice', name: 'W-shape', what: 'Section from the AISC catalogue; its properties drive every check.' },
+          { kind: 'choice', name: 'Steel grade', what: 'A36 or A572 Gr50 / A992 — sets Fy and Fu.' },
           { kind: 'field', name: 'Span L', unit: 'm', what: 'Beam span, over which the uniform loads below are carried.' },
           { kind: 'field', name: 'Unbraced Lb', unit: 'm', what: 'Distance between lateral braces. Compared against Lp and Lr to place the beam in the plastic, inelastic-LTB or elastic-LTB zone.' },
           { kind: 'field', name: 'Cb (moment gradient)', what: 'Lateral-torsional buckling modification factor. 1.0 is conservative.' },
@@ -260,22 +262,61 @@ export const DESIGN_TOOLS: DocTool[] = [
         ],
       },
       {
-        id: 'sd-col',
-        title: 'Column',
+        id: 'sb-out',
+        title: 'Results',
         controls: [
+          { kind: 'output', name: 'φMn and the LTB zone', what: 'Design flexural strength, with a chip saying whether the unbraced length puts it in the plastic, inelastic or elastic range.' },
+          { kind: 'output', name: 'φVn', what: 'Design shear strength, with the web shear coefficient Cv1.' },
+          { kind: 'output', name: 'Deflections', what: 'Live and total service deflection against the L/360 and L/240 limits.' },
+        ],
+      },
+    ],
+  },
+  {
+    id: 'steel-column',
+    name: 'Steel Column',
+    route: '/steel/column',
+    group: 'Steel & timber',
+    summary: 'Flexural buckling, weak-axis flexure, and the combined axial-plus-bending interaction.',
+    basis: 'AISC 360-16 §E3 compression, §H1-1 combined forces.',
+    sections: [
+      {
+        id: 'sc-in',
+        title: 'Inputs',
+        controls: [
+          { kind: 'choice', name: 'W-shape', what: 'Section from the AISC catalogue.' },
+          { kind: 'choice', name: 'Steel grade', what: 'A36 or A572 Gr50 / A992 — sets Fy and Fu.' },
           { kind: 'field', name: 'Height L', unit: 'm', what: 'Unbraced height of the column between restraints.' },
           { kind: 'field', name: 'Kx / Ky', what: 'Effective-length factors about each axis; the larger KL/r governs.' },
-          { kind: 'field', name: 'Dead D / Live L / Pu', unit: 'kN', what: 'Axial demand, entered as service loads or directly as Pu.' },
           { kind: 'choice', name: 'Axial input', what: 'Switches between service-load entry and a direct factored Pu.' },
+          { kind: 'field', name: 'Dead D / Live L / Pu', unit: 'kN', what: 'Axial demand, entered as service loads or directly as Pu.' },
           { kind: 'field', name: 'Mux (strong-axis) / Muy (weak-axis)', unit: 'kN·m', what: 'Moments for the §H1-1 combined check.' },
         ],
       },
       {
-        id: 'sd-conn',
-        title: 'Connections',
+        id: 'sc-out',
+        title: 'Results',
+        controls: [
+          { kind: 'output', name: 'φPn', what: 'Design compressive strength, with the governing slenderness KL/r and the elastic buckling stress Fe.' },
+          { kind: 'output', name: 'Combined ratio §H1-1', what: 'The interaction value, and which of the two equations governed.' },
+        ],
+      },
+    ],
+  },
+  {
+    id: 'bolted-connection',
+    name: 'Bolted Connection',
+    route: '/bolted-connection',
+    group: 'Steel & timber',
+    summary: 'Eccentrically-loaded bolt group by the elastic method, with block shear, out-of-plane tension and prying.',
+    basis: 'AISC 360-16 §J3.6 bolt shear, §J3.10 bearing, §J4.3 block shear, §J3.7 combined tension and shear, §J3.9 prying.',
+    sections: [
+      {
+        id: 'bc-in',
+        title: 'Inputs',
         controls: [
           { kind: 'field', name: 'Applied Vu / Applied Hu (horizontal)', unit: 'kN', what: 'Shear and horizontal demand on the connection.' },
-          { kind: 'choice', name: 'Type / Grade', what: 'Bolt type and grade (A325M, A490M).' },
+          { kind: 'choice', name: 'Grade', what: 'Bolt grade (A325M, A490M), which fixes the nominal shear and tensile strengths.' },
           { kind: 'toggle', name: 'Threads in shear plane (N)', what: 'Threads included in the shear plane reduces the bolt shear strength.' },
           { kind: 'field', name: 'Diameter db', unit: 'mm', what: 'Bolt diameter, which sets the shear, bearing and tear-out strengths.' },
           { kind: 'field', name: 'Rows (vertical) nR / Cols (horizontal) nC', what: 'Bolt rows and columns, which set the group geometry and its polar moment.' },
@@ -284,9 +325,16 @@ export const DESIGN_TOOLS: DocTool[] = [
           { kind: 'field', name: 'In-plane e_x / e_y / Out-of-plane e_out', unit: 'mm', what: 'Load eccentricities. In-plane drives the elastic bolt-group analysis; out-of-plane drives tension and prying.' },
           { kind: 'field', name: 'Gage b (bolt CL → web face)', unit: 'mm', what: 'Distance from the bolt centreline to the web face — the prying lever arm.' },
           { kind: 'field', name: 'Plate thickness t / Plate Fy / Plate Fu', what: 'Shear-tab plate properties.' },
-          { kind: 'field', name: 'Weld size w', unit: 'mm', what: 'Fillet weld leg size, which sets the throat and hence the strength.' },
-          { kind: 'choice', name: 'Electrode', what: 'Filler metal classification, which sets the weld strength.' },
-          { kind: 'choice', name: 'Steel grade', what: 'Member steel grade (Fy/Fu).' },
+        ],
+      },
+      {
+        id: 'bc-out',
+        title: 'Results',
+        controls: [
+          { kind: 'output', name: 'φRn per bolt', what: 'Shear and bearing capacities, and which of the two governs.' },
+          { kind: 'output', name: 'Critical bolt', what: 'The most-loaded bolt from the direct shear plus the torsional component, and its utilisation.' },
+          { kind: 'output', name: 'Block shear §J4.3', what: 'Each tear-out path on the shear tab, with the governing one named.' },
+          { kind: 'output', name: 'Bolt-group layout and 3D scene', what: 'The pattern with the resultant on each bolt, and the tab, bolts and web stub in 3D.' },
         ],
       },
     ],
