--- webapp/src/pages/SteelDesign.tsx
@@ -1,690 +1,0 @@
-import { lazy, Suspense, useMemo, useState } from 'react'
-import { shapesOf, shapeByName } from '../engine/aiscSections'
-import type { AiscShape } from '../engine/aiscSections'
-// type-only imports — no engine code bundled into the browser from here
-import type { BoltGrade, ElectrodeClass } from '../engine/steelDesign'
-import { calcBeam, calcColumn, calcConnection } from '../lib/calcApi'
-import type { BeamCalcResult, ColumnCalcResult, ConnectionCalcResult } from '../lib/calcApi'
-import { useCalcResult } from '../lib/useCalcResult'
-import { Num, Pick, Card, ResultCard, Row } from '../components/qty'
-import { ReportControls } from '../components/ReportControls'
-import { WorkedSolution } from '../components/WorkedSolution'
-import { ConnectionDrawing } from '../components/ConnectionDrawing'
-import type { SolutionStep } from '../lib/solution'
-import { f1, f2, f3 } from '../lib/format'
-import { sn1, sn2, sn3 } from '../lib/solution'
-import { PageHeader } from '../components/calc'
-
-const BeamViewer3D       = lazy(() => import('../components/SteelViewer3D').then(m => ({ default: m.BeamViewer3D })))
-const ColumnViewer3D     = lazy(() => import('../components/SteelViewer3D').then(m => ({ default: m.ColumnViewer3D })))
-const ConnectionViewer3D = lazy(() => import('../components/SteelViewer3D').then(m => ({ default: m.ConnectionViewer3D })))
-
-type Tab      = 'beam' | 'column' | 'connection'
-type Grade    = 'A36' | 'A572G50'
-type ConnType = 'bolt' | 'weld'
-
-const GRADES: Record<Grade, { Fy: number; Fu: number; label: string }> = {
-  A36:     { Fy: 248, Fu: 400, label: 'A36 / SS400  (Fy 248 MPa)' },
-  A572G50: { Fy: 345, Fu: 448, label: 'A572 Gr50 / A992  (Fy 345 MPa)' },
-}
-const W_SHAPES = shapesOf('W')
-
-function shapeOrFirst(name: string): AiscShape { return shapeByName(name) ?? W_SHAPES[0] }
-
-const ok = (pass: boolean, v: string) => (
-  <span className={pass ? 'font-semibold text-green-700' : 'font-semibold text-red-700'}>{v} {pass ? '✓' : '✗'}</span>
-)
-const badge = (zone: string) => {
-  const cls = zone === 'plastic' ? 'bg-green-100 text-green-800' : zone === 'inelastic' ? 'bg-amber-100 text-amber-800' : 'bg-red-100 text-red-800'
-  return <span className={`ml-1.5 rounded px-1.5 py-0.5 text-[10px] font-bold uppercase ${cls}`}>{zone}</span>
-}
-const Spinner    = () => <div className="flex h-72 items-center justify-center rounded-xl border border-slate-200 bg-slate-900 text-sm text-slate-400">Loading 3D…</div>
-const CalcBadge  = ({ loading, error }: { loading: boolean; error: string | null }) => {
-  if (error)   return <span className="ml-2 rounded bg-red-100 px-2 py-0.5 text-xs text-red-700">API error — check console</span>
-  if (loading) return <span className="ml-2 rounded bg-slate-100 px-2 py-0.5 text-xs text-slate-500">computing…</span>
-  return null
-}
-const ShapePick = ({ value, onChange }: { value: string; onChange: (v: string) => void }) => (
-  <label className="col-span-full flex flex-col text-sm">
-    <span className="mb-1 font-medium text-slate-600">W-shape</span>
-    <select value={value} onChange={e => onChange(e.target.value)}
-      className="rounded-md border border-slate-300 px-2.5 py-1.5 text-slate-800 focus:border-[#0056b3] focus:outline-none">
-      {W_SHAPES.map(s => <option key={s.name} value={s.name}>{s.name}</option>)}
-    </select>
-  </label>
-)
-
-// ─── Beam Tab ─────────────────────────────────────────────────────────────
-
-function BeamTab() {
-  const [shapeName, setShapeName] = useState('W310x38.7')
-  const [grade, setGrade]         = useState<Grade>('A572G50')
-  const [span,  setSpan]          = useState(6)
-  const [Lb,    setLb]            = useState(2)
-  const [Cb,    setCb]            = useState(1.0)
-  const [wD,    setWD]            = useState(15)
-  const [wL,    setWL]            = useState(25)
-
-  const { Fy } = GRADES[grade]
-  // Shape geometry stays client-side: needed by 3D viewer and section dimensions in steps.
-  const shape = useMemo(() => shapeOrFirst(shapeName), [shapeName])
-
-  const input = useMemo(
-    () => ({ shapeName, Fy, span, Lb, Cb, wDead: wD, wLive: wL }),
-    [shapeName, Fy, span, Lb, Cb, wD, wL]
-  )
-  const { data: res, loading, error } = useCalcResult<BeamCalcResult>(
-    () => calcBeam(input), [input]
-  )
-
-  const utilM = res ? res.loads.Mu / res.flex.phiMn : 0
-  const utilV = res ? res.loads.Vu / res.shear.phiVn : 0
-
-  const steps = useMemo((): SolutionStep[] => {
-    if (!res) return []
-    const { props, flex, shear, loads } = res
-    const E = 200000
-    return [
-      {
-        title: 'Factored loads (NSCP/AISC load combos)',
-        lines: [
-          { tex: `w_u = \\max(1.4 \\times ${sn1(wD)},\\; 1.2 \\times ${sn1(wD)} + 1.6 \\times ${sn1(wL)}) = ${sn1(loads.wu)}\\text{ kN/m}` },
-          { tex: `M_u = \\frac{w_u L^2}{8} = \\frac{${sn1(loads.wu)} \\times ${sn1(span)}^2}{8} = ${sn1(loads.Mu)}\\text{ kN·m}` },
-          { tex: `V_u = \\frac{w_u L}{2} = ${sn1(loads.Vu)}\\text{ kN}` },
-        ],
-      },
-      {
-        title: 'Compact section (Table B4.1b)',
-        lines: [
-          { tex: `\\lambda_f = \\frac{b_f}{2t_f} = \\frac{${shape.bf}}{2 \\times ${shape.tf}} = ${sn2(flex.lambdaF)}` },
-          { tex: `\\lambda_{pf} = 0.38\\sqrt{E/F_y} = 0.38\\sqrt{${E}/${Fy}} = ${sn2(flex.lambdaPF)}\\quad ${flex.compactFlange ? '\\checkmark\\text{ compact}' : '\\text{non-compact}'}` },
-          { tex: `\\lambda_w = h_w / t_w = ${sn1(props.hw)} / ${shape.tw} = ${sn1(flex.lambdaW)}\\quad \\lambda_{pw} = ${sn1(flex.lambdaPW)}\\quad ${flex.compactWeb ? '\\checkmark' : '\\times'}` },
-        ],
-      },
-      {
-        title: 'Flexural capacity §F2 — lateral-torsional buckling',
-        lines: [
-          { tex: `M_p = F_y Z_x = ${Fy} \\times ${(props.Zx / 1000).toFixed(0)} \\times 10^3\\text{ mm}^3 = ${sn1(flex.Mp)}\\text{ kN·m}` },
-          { tex: `L_p = 1.76\\, r_y \\sqrt{E/F_y} = 1.76 \\times ${shape.ry} \\times \\sqrt{${E}/${Fy}} = ${sn1(flex.Lp)}\\text{ mm} = ${sn2(flex.Lp/1000)}\\text{ m}` },
-          { tex: `L_r = ${sn1(flex.Lr)}\\text{ mm} = ${sn2(flex.Lr/1000)}\\text{ m}` },
-          { text: `L_b = ${Lb} m → zone: ${flex.ltbZone.toUpperCase()}` },
-          { tex: `M_n = ${sn1(flex.Mn)}\\text{ kN·m}\\quad \\phi M_n = 0.90 \\times ${sn1(flex.Mn)} = ${sn1(flex.phiMn)}\\text{ kN·m}` },
-          { tex: `\\text{Utilisation} = \\frac{M_u}{\\phi M_n} = \\frac{${sn1(loads.Mu)}}{${sn1(flex.phiMn)}} = ${sn2(utilM)}\\quad ${utilM <= 1 ? '\\checkmark' : '\\times'}` },
-        ],
-      },
-      {
-        title: 'Shear capacity §G2.1',
-        lines: [
-          { tex: `A_w = d \\cdot t_w = ${shape.d} \\times ${shape.tw} = ${shear.Aw.toFixed(0)}\\text{ mm}^2` },
-          { tex: `h/t_w = ${sn1(shear.hwTw)}\\quad 2.24\\sqrt{E/F_y} = ${sn1(2.24 * Math.sqrt(200000 / Fy))}\\quad C_{v1} = ${sn2(shear.Cv1)},\\; \\phi_v = ${shear.phiV}` },
-          { tex: `\\phi V_n = \\phi_v \\cdot 0.6 F_y A_w C_{v1} = ${sn2(shear.phiVn)}\\text{ kN}` },
-        ],
-      },
-      {
-        title: 'Deflection (unfactored service loads)',
-        lines: [
-          { tex: `\\delta = \\frac{5wL^4}{384EI}` },
-          { tex: `\\delta_L = ${sn2(loads.deltaL)}\\text{ mm}\\quad L/360 = ${sn2(loads.limL360)}\\text{ mm}\\quad ${loads.deltaL <= loads.limL360 ? '\\checkmark' : '\\times EXCEEDS'}` },
-          { tex: `\\delta_{D+L} = ${sn2(loads.deltaD + loads.deltaL)}\\text{ mm}\\quad L/240 = ${sn2(loads.limL240)}\\text{ mm}\\quad ${loads.deltaD + loads.deltaL <= loads.limL240 ? '\\checkmark' : '\\times EXCEEDS'}` },
-        ],
-      },
-    ]
-  }, [res, shape, Fy, wD, wL, span, Lb, utilM])
-
-  return (
-    <div>
-      <div className="grid grid-cols-1 gap-6 lg:grid-cols-[1.4fr_1fr]">
-      <div className="space-y-5">
-        <Card title={<>Section & grade<CalcBadge loading={loading} error={error} /></>}>
-          <ShapePick value={shapeName} onChange={setShapeName} />
-          <Pick label="Steel grade" value={grade} onChange={v => setGrade(v as Grade)}
-            options={Object.entries(GRADES).map(([k, v]) => [k as Grade, v.label])} />
-        </Card>
-        <Card title="Span & bracing">
-          <Num label="Span L" unit="m" value={span} onChange={setSpan} />
-          <Num label="Unbraced Lb" unit="m" value={Lb} onChange={setLb} />
-          <Num label="Cb (moment gradient)" value={Cb} onChange={setCb} />
-        </Card>
-        <Card title="Uniform service loads">
-          <Num label="Dead wD" unit="kN/m" value={wD} onChange={setWD} />
-          <Num label="Live wL" unit="kN/m" value={wL} onChange={setWL} />
-        </Card>
-      </div>
-
-      <div className="space-y-4 lg:sticky lg:top-4 lg:max-h-[calc(100vh-2rem)] lg:self-start lg:overflow-y-auto lg:pr-1">
-        <Suspense fallback={<Spinner />}>
-          <BeamViewer3D shape={shape} span={span} wDead={wD} wLive={wL} />
-        </Suspense>
-
-        {res && (<>
-          <ResultCard title="Section properties">
-            <Row label="A"  value={`${shape.A.toLocaleString()} mm²`} />
-            <Row label="Ix" value={`${(res.props.Ix/1e6).toFixed(1)} ×10⁶ mm⁴`} />
-            <Row label="Sx" value={`${(res.props.Sx/1e3).toFixed(0)} ×10³ mm³`} />
-            <Row label="Zx" value={`${(res.props.Zx/1e3).toFixed(0)} ×10³ mm³`} />
-            <Row label="Lp / Lr" value={`${f2(res.flex.Lp/1000)} / ${f2(res.flex.Lr/1000)} m`} />
-          </ResultCard>
-          <ResultCard title={<>Flexure §F2 {badge(res.flex.ltbZone)}</>}>
-            <Row label="φMn" value={`${f1(res.flex.phiMn)} kN·m`} />
-            <Row alert={utilM>1} label="Mu / φMn" value={ok(utilM<=1, `${(utilM*100).toFixed(0)} %`)} />
-          </ResultCard>
-          <ResultCard title="Shear §G2.1">
-            <Row label="φVn" value={`${f1(res.shear.phiVn)} kN`} sub={`Cv1=${res.shear.Cv1.toFixed(2)}`} />
-            <Row alert={utilV>1} label="Vu / φVn" value={ok(utilV<=1, `${(utilV*100).toFixed(0)} %`)} />
-          </ResultCard>
-          <ResultCard title="Deflection">
-            <Row alert={res.loads.deltaL>res.loads.limL360} label="δL ≤ L/360" value={ok(res.loads.deltaL<=res.loads.limL360, `${f2(res.loads.deltaL)} mm`)} />
-            <Row alert={res.loads.deltaD+res.loads.deltaL>res.loads.limL240} label="δtotal ≤ L/240" value={ok(res.loads.deltaD+res.loads.deltaL<=res.loads.limL240, `${f2(res.loads.deltaD+res.loads.deltaL)} mm`)} />
-          </ResultCard>
-        </>)}
-      </div>
-
-      </div>
-
-      {res && (
-        <div>
-          <WorkedSolution steps={steps} title="Beam Design — step-by-step (AISC 360-16 §F, §G)" />
-        </div>
-      )}
-    </div>
-  )
-}
-
-// ─── Column Tab ────────────────────────────────────────────────────────────
-
-function ColumnTab() {
-  const [shapeName, setShapeName] = useState('W250x67')
-  const [grade, setGrade]         = useState<Grade>('A572G50')
-  const [L,     setL]             = useState(4)
-  const [Kx,    setKx]            = useState(1.0)
-  const [Ky,    setKy]            = useState(1.0)
-  const [dlMode, setDlMode]       = useState<'DL'|'direct'>('DL')
-  const [dead,  setDead]          = useState(800)
-  const [live,  setLive]          = useState(500)
-  const [PuDir, setPuDir]         = useState(1944)
-  const [Mux,   setMux]           = useState(80)
-  const [Muy,   setMuy]           = useState(0)
-
-  const { Fy } = GRADES[grade]
-  const shape  = useMemo(() => shapeOrFirst(shapeName), [shapeName])
-  // Factor axial demand client-side (trivial arithmetic, not proprietary)
-  const Pu     = dlMode === 'DL' ? Math.max(1.4*dead, 1.2*dead+1.6*live) : PuDir
-
-  const input = useMemo(
-    () => ({ shapeName, Fy, L, Kx, Ky, Pu, Mux, Muy }),
-    [shapeName, Fy, L, Kx, Ky, Pu, Mux, Muy]
-  )
-  const { data: res, loading, error } = useCalcResult<ColumnCalcResult>(
-    () => calcColumn(input), [input]
-  )
-
-  const steps = useMemo((): SolutionStep[] => {
-    if (!res) return []
-    const { axial, flexX, weak, comb } = res
-    const E = 200000
-    return [
-      {
-        title: 'Factored axial load',
-        lines: dlMode === 'DL' ? [
-          { tex: `P_u = \\max(1.4D,\\;1.2D+1.6L) = \\max(${sn1(1.4*dead)},\\;${sn1(1.2*dead+1.6*live)}) = ${sn1(Pu)}\\text{ kN}` },
-        ] : [{ tex: `P_u = ${sn1(Pu)}\\text{ kN (direct input)}` }],
-      },
-      {
-        title: 'Slenderness and critical stress §E3',
-        lines: [
-          { tex: `KL/r_x = \\frac{${Kx}\\times${L*1000}}{${shape.rx}} = ${sn1(axial.slendernessX)}` },
-          { tex: `KL/r_y = \\frac{${Ky}\\times${L*1000}}{${shape.ry}} = ${sn1(axial.slendernessY)}\\quad \\text{(governing = }${sn1(axial.slenderness)})` },
-          { tex: `4.71\\sqrt{E/F_y} = 4.71\\sqrt{${E}/${Fy}} = ${sn1(4.71*Math.sqrt(E/Fy))}` },
-          { tex: `F_{cr} = ${sn2(axial.Fcr)}\\text{ MPa}\\quad \\phi P_n = 0.90 \\times ${sn2(axial.Fcr)} \\times ${shape.A} / 1000 = ${sn1(axial.phiPn)}\\text{ kN}` },
-        ],
-        note: axial.slenderOK ? undefined : 'KL/r > 200 — slenderness exceeds §E2 advisory limit.',
-      },
-      {
-        title: 'Flexural capacities',
-        lines: [
-          { tex: `\\phi M_{nx} = ${sn1(flexX.phiMn)}\\text{ kN·m}\\quad (\\text{${flexX.ltbZone} zone, §F2})` },
-          { tex: `\\phi M_{ny} = 0.90 F_y Z_y = ${sn1(weak.phiMny)}\\text{ kN·m}\\quad (\\text{§F6, weak-axis})` },
-        ],
-      },
-      {
-        title: `Combined loading §H1-1 (equation ${comb.equation})`,
-        lines: comb.equation === 'H1-1a' ? [
-          { tex: `P_u/\\phi P_n = ${sn3(Pu/axial.phiPn)} \\geq 0.2 \\Rightarrow \\text{use §H1-1a}` },
-          { tex: `\\frac{P_u}{\\phi P_n} + \\frac{8}{9}\\!\\left(\\frac{M_{ux}}{\\phi M_{nx}} + \\frac{M_{uy}}{\\phi M_{ny}}\\right) = ${sn3(Pu/axial.phiPn)} + \\frac{8}{9}(${sn3(Mux/flexX.phiMn)} + ${sn3(Muy > 0 ? Muy/weak.phiMny : 0)}) = ${sn3(comb.ratio)}\\quad${comb.ok?'\\checkmark':'\\times'}` },
-        ] : [
-          { tex: `P_u/\\phi P_n = ${sn3(Pu/axial.phiPn)} < 0.2 \\Rightarrow \\text{use §H1-1b}` },
-          { tex: `\\frac{P_u}{2\\phi P_n} + \\left(\\frac{M_{ux}}{\\phi M_{nx}} + \\frac{M_{uy}}{\\phi M_{ny}}\\right) = ${sn3(comb.ratio)}\\quad${comb.ok?'\\checkmark':'\\times'}` },
-        ],
-      },
-    ]
-  }, [res, shape, Pu, Mux, Muy, Kx, Ky, L, Fy, dead, live, dlMode])
-
-  return (
-    <div>
-      <div className="grid grid-cols-1 gap-6 lg:grid-cols-[1.4fr_1fr]">
-      <div className="space-y-5">
-        <Card title={<>Section & grade<CalcBadge loading={loading} error={error} /></>}>
-          <ShapePick value={shapeName} onChange={setShapeName} />
-          <Pick label="Steel grade" value={grade} onChange={v => setGrade(v as Grade)}
-            options={Object.entries(GRADES).map(([k, v]) => [k as Grade, v.label])} />
-        </Card>
-        <Card title="Column geometry">
-          <Num label="Height L" unit="m" value={L} onChange={setL} />
-          <Num label="Kx" value={Kx} onChange={setKx} />
-          <Num label="Ky" value={Ky} onChange={setKy} />
-        </Card>
-        <Card title="Loads">
-          <Pick label="Axial input" value={dlMode} onChange={v => setDlMode(v as 'DL'|'direct')}
-            options={[['DL','Service D & L'],['direct','Factored Pu']]} />
-          {dlMode === 'DL' ? <>
-            <Num label="Dead D" unit="kN" value={dead} onChange={setDead} />
-            <Num label="Live L" unit="kN" value={live} onChange={setLive} />
-          </> : <Num label="Pu" unit="kN" value={PuDir} onChange={setPuDir} />}
-          <Num label="Mux (strong-axis)" unit="kN·m" value={Mux} onChange={setMux} />
-          <Num label="Muy (weak-axis)" unit="kN·m" value={Muy} onChange={setMuy} />
-        </Card>
-      </div>
-
-      <div className="space-y-4 lg:sticky lg:top-4 lg:max-h-[calc(100vh-2rem)] lg:self-start lg:overflow-y-auto lg:pr-1">
-        <Suspense fallback={<Spinner />}>
-          <ColumnViewer3D shape={shape} L={L} Pu={Pu} Mux={Mux} />
-        </Suspense>
-        {res && (<>
-          <ResultCard title="Axial §E3">
-            <Row label="KL/rx" value={f1(res.axial.slendernessX)} />
-            <Row label="KL/ry" value={f1(res.axial.slendernessY)} sub="governing" />
-            <Row alert={!res.axial.slenderOK} label="KL/r" value={`${f1(res.axial.slenderness)}${res.axial.slenderOK?'':' > 200 ✗'}`} />
-            <Row label="Fcr" value={`${f1(res.axial.Fcr)} MPa`} />
-            <Row label="φPn" value={`${f1(res.axial.phiPn)} kN`} />
-          </ResultCard>
-          <ResultCard title="Flexure">
-            <Row label="φMnx" value={`${f1(res.flexX.phiMn)} kN·m`} sub={res.flexX.ltbZone} />
-            <Row label="φMny §F6" value={`${f1(res.weak.phiMny)} kN·m`} />
-          </ResultCard>
-          <ResultCard title={`Combined §${res.comb.equation}`}>
-            <Row alert={!res.comb.ok} label="Combined ratio" value={ok(res.comb.ok, `${(res.comb.ratio*100).toFixed(0)} %`)} />
-          </ResultCard>
-        </>)}
-      </div>
-
-      </div>
-
-      {res && (
-        <div>
-          <WorkedSolution steps={steps} title="Column Design — step-by-step (AISC 360-16 §E3, §H1-1)" />
-        </div>
-      )}
-    </div>
-  )
-}
-
-// ─── Connection Tab ─────────────────────────────────────────────────────────
-
-function ConnectionTab() {
-  const [connType,   setConnType]   = useState<ConnType>('bolt')
-  const [Vu,         setVu]         = useState(150)
-  const [Hu,         setHu]         = useState(0)
-  const [boltGrade,  setBoltGrade]  = useState<BoltGrade>('A325M')
-  const [db,         setDb]         = useState(20)
-  const [nRows,      setNRows]      = useState(3)
-  const [nCols,      setNCols]      = useState(1)
-  const [sy,         setSy]         = useState(70)
-  const [sx,         setSx]         = useState(70)
-  const [ey,         setEy]         = useState(40)
-  const [ex_edge,    setExEdge]     = useState(35)
-  const [threads,    setThreads]    = useState<'yes'|'no'>('yes')
-  const [tPlate,     setTPlate]     = useState(10)
-  const [FuPlate,    setFuPlate]    = useState(400)
-  const [FyPlate,    setFyPlate]    = useState(248)
-  const [ex_load,    setExLoad]     = useState(0)
-  const [ey_load,    setEyLoad]     = useState(0)
-  const [e_out,      setEOut]       = useState(0)
-  const [b_gage,     setBGage]      = useState(0)
-  const [electrode,  setElectrode]  = useState<ElectrodeClass>('E70')
-  const [wSize,      setWSize]      = useState(8)
-
-  const input = useMemo(
-    () => ({
-      Vu, Hu,
-      boltGrade, db, nRows, nCols, sy, sx, ey, ex_edge,
-      threads: threads === 'yes',
-      tPlate, FuPlate, FyPlate,
-      ex_load, ey_load, e_out, b_gage,
-      electrode, wSize,
-    }),
-    [Vu, Hu, boltGrade, db, nRows, nCols, sy, sx, ey, ex_edge, threads,
-     tPlate, FuPlate, FyPlate, ex_load, ey_load, e_out, b_gage, electrode, wSize]
-  )
-  const { data: res, loading, error } = useCalcResult<ConnectionCalcResult>(
-    () => calcConnection(input), [input]
-  )
-
-  const govBlockShear = res
-    ? res.blockShear.reduce((mn, c) => c.phiRn < mn.phiRn ? c : mn, res.blockShear[0])
-    : null
-
-  const boltSteps = useMemo((): SolutionStep[] => {
-    if (!res) return []
-    const { phiRnBolt, geom, eccentric, outOfPlane, prying, blockShear } = res
-    const { Fnv, Ab, phiRn_shear, phiRn_bearing, phiRn } = phiRnBolt
-    const { n, Ip } = geom
-    const M = eccentric.M
-    return [
-      {
-        title: `Bolt shear capacity §J3.6 — ${boltGrade}, d_b = ${db} mm`,
-        lines: [
-          { tex: `A_b = \\frac{\\pi}{4} d_b^2 = \\frac{\\pi}{4}(${db})^2 = ${Ab.toFixed(0)}\\text{ mm}^2` },
-          { tex: `F_{nv} = ${Fnv}\\text{ MPa}\\quad (${threads==='yes'?'threads in shear plane, N':'threads excluded, X'})` },
-          { tex: `\\phi R_{n,\\text{shear}} = 0.75 F_{nv} A_b = 0.75 \\times ${Fnv} \\times ${Ab.toFixed(0)} / 1000 = ${sn2(phiRn_shear)}\\text{ kN/bolt}` },
-        ],
-      },
-      {
-        title: `Bearing on plate §J3.10 — t = ${tPlate} mm, F_u = ${FuPlate} MPa`,
-        lines: [
-          { tex: `\\phi R_{n,\\text{br}} = 0.75 \\times 2.4 F_u d_b t = 0.75 \\times 2.4 \\times ${FuPlate} \\times ${db} \\times ${tPlate} / 1000 = ${sn2(phiRn_bearing)}\\text{ kN/bolt}` },
-          { tex: `\\phi R_n\\text{ (governing)} = \\min(${sn2(phiRn_shear)},\\;${sn2(phiRn_bearing)}) = ${sn2(phiRn)}\\text{ kN/bolt}` },
-        ],
-      },
-      {
-        title: `Bolt group (${nRows} × ${nCols} = ${n} bolts) — in-plane eccentricity (elastic method)`,
-        lines: [
-          { tex: `I_p = \\sum(x_i^2 + y_i^2) = ${Ip.toFixed(0)}\\text{ mm}^2` },
-          { tex: `M = V_u \\cdot e_x - H_u \\cdot e_y = ${Vu} \\times ${ex_load} - ${Hu} \\times ${ey_load} = ${M.toFixed(0)}\\text{ kN·mm}` },
-          { text: `Direct shear per bolt: Vx = Hu/n = ${(Hu/n).toFixed(2)} kN,  Vy = Vu/n = ${(Vu/n).toFixed(2)} kN` },
-          { tex: `V_{x,i} = H_u/n - M y_i / I_p\\quad V_{y,i} = V_u/n + M x_i / I_p` },
-          { tex: `R_{\\max} = ${sn2(eccentric.Rmax)}\\text{ kN on bolt }\\textit{${eccentric.critical}}` },
-          { tex: `\\text{Utilisation} = R_{\\max} / (\\phi R_n) = ${sn2(eccentric.Rmax)} / ${sn2(phiRn)} = ${sn2(eccentric.Rmax/phiRn)}\\quad ${eccentric.Rmax<=phiRn?'\\checkmark':'\\times'}` },
-        ],
-      },
-      {
-        title: 'Bearing stress and shear stress on critical bolt',
-        lines: (() => {
-          const crit = eccentric.bolts.find(b => b.id === eccentric.critical)
-          if (!crit) return []
-          const Ab2 = (Math.PI/4)*db*db
-          return [
-            { tex: `f_{br} = \\frac{R_{\\max}}{d_b \\cdot t} = \\frac{${sn2(eccentric.Rmax)} \\times 1000}{${db} \\times ${tPlate}} = ${sn1(crit.fbr)}\\text{ MPa}\\quad \\phi F_{br} = 0.75 \\times 2.4 \\times ${FuPlate} = ${sn1(0.75*2.4*FuPlate)}\\text{ MPa}\\quad ${crit.fbr <= 0.75*2.4*FuPlate ? '\\checkmark':'\\times'}` },
-            { tex: `f_v = \\frac{R_{\\max}}{A_b} = \\frac{${sn2(eccentric.Rmax)} \\times 1000}{${Ab2.toFixed(0)}} = ${sn1(crit.fv)}\\text{ MPa}\\quad \\phi F_{nv} = 0.75 \\times ${Fnv} = ${sn1(0.75*Fnv)}\\text{ MPa}\\quad ${crit.fv <= 0.75*Fnv ? '\\checkmark':'\\times'}` },
-          ]
-        })(),
-      },
-      ...(outOfPlane ? [{
-        title: `Out-of-plane eccentricity §J3.7 — e_out = ${e_out} mm`,
-        lines: [
-          { tex: `M_{op} = V_u \\cdot e_{out} = ${Vu} \\times ${e_out} = ${outOfPlane.M_op.toFixed(0)}\\text{ kN·mm}` },
-          { tex: `\\text{Neutral axis at bottom bolt. }\\sum y_i^2 = ${outOfPlane.sumYi2.toFixed(0)}\\text{ mm}^2` },
-          { tex: `T_i = \\frac{M_{op} \\cdot y_i}{\\sum y_i^2}\\quad (y_i \\text{ measured from lowest bolt})` },
-          { tex: `T_{\\max} = ${f2(outOfPlane.Tmax)}\\text{ kN on bolt }\\textit{${outOfPlane.critical}}` },
-          { tex: `\\text{AISC Table J3.2: } F_{nt} = ${outOfPlane.Fnt}\\text{ MPa},\\; F_{nv} = ${outOfPlane.Fnv}\\text{ MPa}` },
-          { tex: `\\phi F_{nt}' = \\min\\!\\left(1.3 F_{nt} - \\frac{F_{nt}}{\\phi F_{nv}} f_{rv},\\; F_{nt}\\right) \\geq 0\\quad \\S J3.7` },
-          { tex: `\\phi T_n = \\phi F_{nt}' A_b / 1000 = ${f2(outOfPlane.phiTn_crit)}\\text{ kN/bolt}` },
-          { tex: `\\text{Utilisation} = T_{\\max} / (\\phi T_n) = ${(outOfPlane.Tmax / outOfPlane.phiTn_crit * 100).toFixed(0)}\\%\\quad ${outOfPlane.ok ? '\\checkmark' : '\\times'}` },
-        ],
-      },
-      ...(prying ? [{
-        title: `Prying action §J3.9 — b = ${b_gage} mm`,
-        lines: [
-          { tex: `b' = b - d_b/2 = ${b_gage} - ${db}/2 = ${prying.b_prime.toFixed(1)}\\text{ mm}` },
-          { tex: `a' = \\min(a,\\; 1.25b) = \\min(${ex_edge},\\; ${(1.25*b_gage).toFixed(1)}) = ${prying.a_prime.toFixed(1)}\\text{ mm}` },
-          { tex: `\\rho = b'/a' = ${prying.rho.toFixed(3)},\\quad \\delta = 1 - d_h/p = 1 - ${db+2}/${sy} = ${prying.delta.toFixed(3)}` },
-          { tex: `\\beta = \\frac{1}{\\rho}\\left(\\frac{\\phi B_n}{T_{\\max}} - 1\\right) = ${prying.beta.toFixed(3)},\\quad \\alpha = ${prying.alpha.toFixed(3)}` },
-          { tex: `Q = \\alpha \\delta \\rho T_{\\max} = ${f2(prying.Q)}\\text{ kN}\\quad\\Rightarrow\\quad T_{\\text{total}} = ${f2(prying.T_total)}\\text{ kN}` },
-          { tex: `t_{\\text{req}} = \\sqrt{\\frac{4 T b'}{\\phi_f F_y p (1+\\delta\\alpha)}} = ${prying.t_req.toFixed(1)}\\text{ mm}\\quad(\\text{plate }t=${tPlate}\\text{ mm}\\quad${tPlate>=prying.t_req?'\\checkmark':'\\times'})` },
-          { tex: `t_0 = \\sqrt{\\frac{4\\phi B_n b'}{\\phi_f F_y p}} = ${prying.t_no_prying.toFixed(1)}\\text{ mm}\\quad(\\text{min for zero prying})` },
-        ],
-      }] : [])] : []),
-      {
-        title: 'Block shear §J4.3 (shear tab, single bolt line)',
-        lines: blockShear.flatMap(c => [
-          { text: c.label },
-          { tex: `A_{gv} = ${c.Agv.toFixed(0)}\\text{ mm}^2\\quad A_{nv} = ${c.Anv.toFixed(0)}\\text{ mm}^2\\quad A_{nt} = ${c.Ant.toFixed(0)}\\text{ mm}^2` },
-          { tex: `R_{n,\\text{fract}} = 0.6 F_u A_{nv} + U_{bs} F_u A_{nt} = ${sn1(c.Rn_fract)}\\text{ kN}` },
-          { tex: `\\text{cap } = 0.6 F_y A_{gv} + U_{bs} F_u A_{nt} = ${sn1(c.Rn_cap)}\\text{ kN}` },
-          { tex: `\\phi R_n = 0.75 \\times ${sn1(Math.min(c.Rn_fract,c.Rn_cap))} = ${sn1(c.phiRn)}\\text{ kN}\\quad ${Vu<=c.phiRn?'\\checkmark':'\\times'}` },
-        ]),
-      },
-    ]
-  }, [res, boltGrade, db, nRows, nCols, sy, ex_edge, tPlate, FuPlate, threads, Vu, Hu, ex_load, ey_load, e_out, b_gage])
-
-  return (
-    <div>
-      <div className="grid grid-cols-1 gap-6 lg:grid-cols-[1.2fr_0.9fr_1fr]">
-      {/* ── inputs ── */}
-      <div className="space-y-5">
-        <Card title={<>Connection<CalcBadge loading={loading} error={error} /></>}>
-          <Pick label="Type" value={connType} onChange={v => setConnType(v as ConnType)}
-            options={[['bolt','Bolted (§J3)'],['weld','Fillet weld (§J2.4)']]} />
-          <Num label="Applied Vu" unit="kN" value={Vu} onChange={setVu} />
-          <Num label="Applied Hu (horizontal)" unit="kN" value={Hu} onChange={setHu} />
-        </Card>
-
-        {connType === 'bolt' ? (<>
-          <Card title="Bolt properties">
-            <Pick label="Grade" value={boltGrade} onChange={v => setBoltGrade(v as BoltGrade)}
-              options={[['A325M','A325M (F10T)'],['A490M','A490M (F13T)']]} />
-            <Num label="Diameter db" unit="mm" value={db} onChange={setDb} />
-            <Pick label="Threads in shear plane" value={threads} onChange={v => setThreads(v as 'yes'|'no')}
-              options={[['yes','Yes (N)'],['no','No (X)']]} />
-          </Card>
-          <Card title="Bolt pattern">
-            <Num label="Rows (vertical) nR" value={nRows} onChange={setNRows} />
-            <Num label="Cols (horizontal) nC" value={nCols} onChange={setNCols} />
-            <Num label="Vertical spacing sv" unit="mm" value={sy} onChange={setSy} />
-            <Num label="Horizontal spacing sh" unit="mm" value={sx} onChange={setSx} />
-            <Num label="Edge dist vertical ey" unit="mm" value={ey} onChange={setEy} />
-            <Num label="Edge dist horiz ex" unit="mm" value={ex_edge} onChange={setExEdge} />
-          </Card>
-          <Card title="Eccentricity (load point from bolt centroid)">
-            <Num label="In-plane e_x" unit="mm" value={ex_load} onChange={setExLoad} />
-            <Num label="In-plane e_y" unit="mm" value={ey_load} onChange={setEyLoad} />
-            <Num label="Out-of-plane e_out" unit="mm" value={e_out} onChange={setEOut} />
-            <p className="col-span-full text-[10px] text-slate-500">
-              e_x/e_y: in-plane offset (§J3.6 elastic method).
-              e_out: perpendicular to plate → bolt tension + §J3.7 interaction.
-            </p>
-          </Card>
-          {e_out > 0 && (
-            <Card title="Prying action §J3.9">
-              <Num label="Gage b (bolt CL → web face)" unit="mm" value={b_gage} onChange={setBGage} />
-              <p className="col-span-full text-[10px] text-slate-500">
-                b = distance from bolt centreline to face of the connecting web or stem.
-                Set 0 to skip prying check. Edge dist a = ex, pitch p = sv, plate tf/Fy reused from above.
-              </p>
-            </Card>
-          )}
-          <Card title="Plate / connected part">
-            <Num label="Plate thickness t" unit="mm" value={tPlate} onChange={setTPlate} />
-            <Num label="Plate Fy" unit="MPa" value={FyPlate} onChange={setFyPlate} />
-            <Num label="Plate Fu" unit="MPa" value={FuPlate} onChange={setFuPlate} />
-          </Card>
-        </>) : (
-          <Card title="Fillet weld">
-            <Pick label="Electrode" value={electrode} onChange={v => setElectrode(v as ElectrodeClass)}
-              options={[['E70','E70XX (482 MPa)'],['E80','E80XX (550 MPa)'],['E90','E90XX (620 MPa)'],['E100','E100XX (690 MPa)']]} />
-            <Num label="Weld size w" unit="mm" value={wSize} onChange={setWSize} />
-            <Card title="Plate (for weld group geometry)">
-              <Num label="Rows nR (for length)" value={nRows} onChange={setNRows} />
-              <Num label="Vert spacing sv" unit="mm" value={sy} onChange={setSy} />
-              <Num label="Edge dist ey" unit="mm" value={ey} onChange={setEy} />
-            </Card>
-          </Card>
-        )}
-      </div>
-
-      {/* ── 2D drawing + 3D ── */}
-      <div className="space-y-4">
-        {res && (
-          <ConnectionDrawing
-            geom={res.geom} db={db}
-            boltForces={connType === 'bolt' ? res.eccentric.bolts : undefined}
-            critical={connType === 'bolt' ? res.eccentric.critical : undefined}
-            Vu={Vu} Hu={Hu} ex_load={res.geom.Cx + ex_load} ey_load={res.geom.Cy + ey_load}
-            connType={connType}
-          />
-        )}
-        {res && (
-          <Suspense fallback={<Spinner />}>
-            <ConnectionViewer3D geom={res.geom} db={db} t_plate={tPlate} critical={connType === 'bolt' ? res.eccentric.critical : ''} />
-          </Suspense>
-        )}
-      </div>
-
-      {/* ── results ── */}
-      <div className="space-y-4 lg:sticky lg:top-4 lg:max-h-[calc(100vh-2rem)] lg:self-start lg:overflow-y-auto lg:pr-1">
-        {res && connType === 'bolt' && (<>
-          <ResultCard title="Bolt capacity / bolt">
-            <Row label="φRn shear" value={`${f2(res.phiRnBolt.phiRn_shear)} kN`} />
-            <Row label="φRn bearing" value={`${f2(res.phiRnBolt.phiRn_bearing)} kN`} />
-            <Row label="φRn governing" value={<b>{f2(res.phiRnBolt.phiRn)} kN</b>} />
-          </ResultCard>
-          <ResultCard title={`Eccentric group (${res.geom.n} bolts)`}>
-            <Row label="Ip" value={`${res.geom.Ip.toFixed(0)} mm²`} />
-            <Row label="In-plane M" value={`${res.eccentric.M.toFixed(0)} kN·mm`} />
-            <Row label="Critical bolt" value={res.eccentric.critical} sub={`R = ${f2(res.eccentric.Rmax)} kN`} />
-            <Row alert={res.eccentric.Rmax > res.phiRnBolt.phiRn}
-              label="Rmax / φRn"
-              value={ok(res.eccentric.Rmax <= res.phiRnBolt.phiRn, `${(res.eccentric.Rmax/res.phiRnBolt.phiRn*100).toFixed(0)} %`)} />
-          </ResultCard>
-          <ResultCard title="Per-bolt forces">
-            <div className="overflow-x-auto">
-              <table className="w-full text-xs">
-                <thead><tr className="text-left text-slate-500">
-                  <th className="pr-2 pb-1">id</th><th className="pr-2 pb-1">Vx kN</th><th className="pr-2 pb-1">Vy kN</th>
-                  <th className="pr-2 pb-1">R kN</th><th className="pb-1">fbr MPa</th>
-                </tr></thead>
-                <tbody>
-                  {res.eccentric.bolts.map(b => (
-                    <tr key={b.id} className={b.id === res.eccentric.critical ? 'font-semibold text-amber-700' : ''}>
-                      <td className="pr-2">{b.id}</td>
-                      <td className="pr-2">{f2(b.Vx)}</td>
-                      <td className="pr-2">{f2(b.Vy)}</td>
-                      <td className="pr-2">{f2(b.R)}</td>
-                      <td>{f1(b.fbr)}</td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </div>
-          </ResultCard>
-          <ResultCard title="Block shear §J4.3">
-            {res.blockShear.map(c => (
-              <Row key={c.label} alert={Vu > c.phiRn}
-                label={<span className="text-[10px]">{c.label.replace('§J4.3 ', '')}</span>}
-                value={ok(Vu <= c.phiRn, `φRn = ${f1(c.phiRn)} kN`)} />
-            ))}
-            {govBlockShear && (
-              <Row alert={Vu > govBlockShear.phiRn}
-                label="Governing block shear"
-                value={ok(Vu <= govBlockShear.phiRn, `${f1(govBlockShear.phiRn)} kN`)} />
-            )}
-          </ResultCard>
-
-          {res.outOfPlane && (
-            <ResultCard title="Out-of-plane eccentricity §J3.7 (critical bolt)">
-              <Row label="M_op = Vu·e_out" value={`${res.outOfPlane.M_op.toFixed(0)} kN·mm`} />
-              <Row label="Σyi²" value={`${res.outOfPlane.sumYi2.toFixed(0)} mm²`} />
-              <Row label="Critical bolt (max T)" value={res.outOfPlane.critical}
-                sub={`T = ${f2(res.outOfPlane.Tmax)} kN`} />
-              <Row label="φTn (reduced)" value={`${f2(res.outOfPlane.phiTn_crit)} kN`}
-                sub={`φFnt' = ${res.outOfPlane.bolts.find(b=>b.id===res.outOfPlane!.critical)?.phiFnt_prime.toFixed(0)} MPa`} />
-              <Row alert={!res.outOfPlane.ok}
-                label="Combined check (§J3.7)"
-                value={ok(res.outOfPlane.ok, res.outOfPlane.ok ? 'PASS' : 'FAIL')} />
-              <div className="col-span-full overflow-x-auto">
-                <table className="mt-1 w-full text-xs">
-                  <thead><tr className="text-left text-slate-500">
-                    <th className="pr-2 pb-1">id</th><th className="pr-2 pb-1">yi mm</th>
-                    <th className="pr-2 pb-1">T kN</th><th className="pr-2 pb-1">frv MPa</th><th className="pb-1">util</th>
-                  </tr></thead>
-                  <tbody>
-                    {res.outOfPlane.bolts.map(b => (
-                      <tr key={b.id} className={b.id === res.outOfPlane!.critical ? 'font-semibold text-amber-700' : ''}>
-                        <td className="pr-2">{b.id}</td>
-                        <td className="pr-2">{b.yi.toFixed(0)}</td>
-                        <td className="pr-2">{f2(b.T)}</td>
-                        <td className="pr-2">{f1(b.frv)}</td>
-                        <td className={b.util > 1 ? 'text-red-600' : ''}>{(b.util*100).toFixed(0)}%</td>
-                      </tr>
-                    ))}
-                  </tbody>
-                </table>
-              </div>
-            </ResultCard>
-          )}
-
-          {res.prying && (
-            <ResultCard title="Prying action §J3.9 (critical bolt)">
-              <Row label="b' = b − db/2" value={`${res.prying.b_prime.toFixed(1)} mm`} />
-              <Row label="a' = min(a, 1.25b)" value={`${res.prying.a_prime.toFixed(1)} mm`} />
-              <Row label="ρ = b'/a'" value={f3(res.prying.rho)} />
-              <Row label="δ = 1 − dh/p" value={f3(res.prying.delta)} />
-              <Row label="β (prying potential)" value={f3(res.prying.beta)} />
-              <Row label="α (prying fraction)" value={f3(res.prying.alpha)} sub="0 = none, 1 = full" />
-              <Row label="Prying force Q" value={`${f2(res.prying.Q)} kN`} />
-              <Row label="T_total = T + Q" value={`${f2(res.prying.T_total)} kN`} />
-              <Row label="Required plate t" value={`${res.prying.t_req.toFixed(1)} mm`}
-                sub={`t_no_prying = ${res.prying.t_no_prying.toFixed(1)} mm`} />
-              <Row alert={!res.prying.ok}
-                label="T_total ≤ φTn (§J3.9)"
-                value={ok(res.prying.ok, res.prying.ok ? 'PASS' : 'FAIL')} />
-            </ResultCard>
-          )}
-        </>)}
-
-        {res && connType === 'weld' && (
-          <ResultCard title="Fillet weld §J2.4">
-            <Row label="φRnw / mm" value={`${f3(res.weld.phiRnw)} kN/mm`} />
-            <Row label="Required length" value={`${f1(res.weld.L_reqd)} mm`} sub={`plate H = ${f1(res.geom.plateH)} mm`} />
-            <Row alert={Vu > res.weldCapacity}
-              label="Vu vs capacity"
-              value={ok(Vu <= res.weldCapacity, `${f1(res.weldCapacity)} kN`)}
-              sub={`2 × ${res.geom.plateH.toFixed(0)} mm`} />
-          </ResultCard>
-        )}
-      </div>
-
-      </div>
-
-      {res && connType === 'bolt' && (
-        <div>
-          <WorkedSolution steps={boltSteps} title="Connection Design — step-by-step (AISC 360-16 §J3, §J4)" />
-        </div>
-      )}
-    </div>
-  )
-}
-
-// ─── Main page ─────────────────────────────────────────────────────────────
-
-export default function SteelDesign() {
-  const [tab, setTab] = useState<Tab>('beam')
-  const tabBtn = (t: Tab, label: string) => (
-    <button type="button" onClick={() => setTab(t)}
-      className={`rounded-md px-4 py-1.5 text-sm font-semibold transition ${tab === t ? 'bg-[#0056b3] text-white shadow' : 'text-slate-600 hover:bg-slate-100'}`}>
-      {label}
-    </button>
-  )
-  return (
-    <div>
-      <PageHeader title="Steel Design" badges={['NSCP 2015', 'ACI 318-14']} />
-      <div className="mx-auto max-w-[1500px] px-6 pb-6 pt-5">
-      <p className="no-print mt-1 text-slate-600">
-        AISC 360-16 LRFD — beams (§F/G + deflections), columns (§E3 + §H1-1), connections
-        (bolt group with in-plane eccentricity, bearing/shear stresses, block shear §J4.3;
-        fillet welds §J2.4). 3D scene + 2D layout + step-by-step solution on every tab.
-      </p>
-      <ReportControls title="Steel Design Report" />
-      <div className="no-print mt-5 mb-6 flex gap-2 rounded-xl border border-slate-200 bg-white p-1.5 shadow-sm w-fit">
-        {tabBtn('beam',       'Beam')}
-        {tabBtn('column',     'Column')}
-        {tabBtn('connection', 'Connection')}
-      </div>
-      <div className="mt-2">
-        {tab === 'beam'       && <BeamTab />}
-        {tab === 'column'     && <ColumnTab />}
-        {tab === 'connection' && <ConnectionTab />}
-      </div>
-      </div>
-    </div>
-  )
-}
